requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.11.0
django-slick-reporting==0.5.5
django-widget-tweaks==1.4.8
python-dateutil==2.8.1
pytz==2020.5
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
